+++ packages/microcosm-preact/src/action-button.js
@@ -1,27 +1,19 @@
 import { h, Component } from 'preact'
 import { Action, merge } from 'microcosm'
+import ActionQueue from 'microcosm/addons/action-queue'
 
 class ActionButton extends Component {
   constructor(props, context) {
     super(props, context)
 
     this.send = this.props.send || this.context.send
-
     this.click = this.click.bind(this)
+
+    this._queue = new ActionQueue(this)
   }
 
-  click(event) {
-    let action = this.send(this.props.action, this.props.value)
-
-    if (action instanceof Action) {
-      action.subscribe(this.props)
-    }
-
-    if (this.props.onClick) {
-      this.props.onClick(event, action)
-    }
-
-    return action
+  componentWillUnmount() {
+    this._queue.empty()
   }
 
   render() {
@@ -42,6 +34,20 @@ class ActionButton extends Component {
     }
 
     return h(this.props.tag, props)
+  }
+
+  click(event) {
+    let action = this.send(this.props.action, this.props.value)
+
+    if (action instanceof Action) {
+      this._queue.push(action, this.props)
+    }
+
+    if (this.props.onClick) {
+      this.props.onClick(event, action)
+    }
+
+    return action
   }
 }