+++ packages/microcosm/src/addons/jest-matchers.js
@@ -37,7 +37,7 @@ expect.extend({
     return {
       pass: pass,
       message: () => {
-        return `Expected action ${operator} be '${status}'.`
+        return `Expected action ${operator} be '${status}'. Instead got ${action.status}.`
       }
     }
   },